Support mobile resizing by window content area
For supported apps (Safari, Google Chrome, and Firefox), resize the
window content area (window body) to fit the mobile sizes. For all other
sizes and for unsupported apps, resize the entire window (window
frame/bounds) to the selected dimensions.

Note that Firefox doesn't provide access to the inner window dimensions,
so a best-guess hardcoded value was used to account for the window
frame. The value is derived from the height of the tab bar and bookmarks
toolbar for the current version of Firefox, so the value could become
outdated with Firefox updates and/or if other toolbars are installed or
any default ones hidden.

Developer note: The Factory Pattern was used to implement the supported
apps.