Workaround FastScripts keyboard shortcut bug
When invoking the script on a Safari window with a control key shortcut
using FastScripts, the script can't access the desired UI element with
GUI scripting. The control key causes a "Display a menu" pop-up label to
appear in the Safari window and it apparently interferes with UI
scripting. FastScripts apparently does not release keys used to invoke a
script with a keyboard shortcut even after a delay, so the workaround
is to add 'key up' commands to the script to get modifier keys to
release. Hopefully, modifier keys are the only troublesome keys because
it would be impractical to script 'key up' commands for every single key
on the keyboard.