Get inner window height with JavaScript
Google Chrome doesn't provide access to the dimensions of the window
scroll area via UI Scripting, so use JavaScript instead.

Also use JavaScript as a fallback for Safari/Webkit in case an app
update breaks GUI scripting.

Use a constructor to create an instance of the factory object because
some properties were getting retained between script runs when it was a
bare script object.