Optimize list access speed as much as possible
Add the standard AppleScript list access speed hacks (by creating
references) in an attempt to speed up the recursive subroutine. List
access doesn't appear to be the source of the slowness though. The
slowdown is probably the result of recursively searching UI elements
because UI scripting is slow in general.