Speed optimizations: Prefer JavaScript over UI scripting
Most people probably have JavaScript enabled in their web browsers, so
try that first since it would totally bypass the need for UI scripting
which is pretty slow.

As a fallback if JavaScript fails, restore the previous UI scripting
algorithm that targets specific UI elements for known/tested versions
of Safari (7-9) on OS X 9 thru 11.

As a last resort if the targeted UI scripting fails, use the new
UI scripting algorithm which recursively searches all UI elements for a
scroll area containing a web area. This is much slower, but it is a more
flexible method that can find the target UI element without knowing in
advance the exact UI element hierarchy. Hopefully, it will handle future
OS and app changes to the Accessibility API that would normally break UI
scripting (as the changes from Safari 7 through 9 did).